# Code optimization (first round)