--- a/R/check_arg.R
+++ b/R/check_arg.R
@@ -115,7 +115,10 @@ check_info <- check_increasing
 #' @param K Number of total analysis.
 #' @noRd
 check_theta <- function(theta, K) {
-  if (length(theta) != K) stop("if length(theta) > 1, must be same as info!")
+  n <- length(theta)
+  if (n > 1 && n != K) stop(
+    "the length of `theta` (if greater than 1) must be same as the number of analyses!"
+  )
   if (theta[K] < 0) stop("final effect size of `theta` must be non-negative!")
 }