--- a/R/check_arg.R
+++ b/R/check_arg.R
@@ -112,32 +112,34 @@ check_info <- check_increasing
 #' Check the argument `theta`
 #'
 #' @param theta Treatment effect.
-#' @param K Number of total analysis.
+#' @param k Number of total analysis.
 #' @noRd
-check_theta <- function(theta, K) {
+check_theta <- function(theta, k) {
   n <- length(theta)
-  if (n > 1 && n != K) stop(
-    "the length of `theta` (if greater than 1) must be same as the number of analyses!"
+  if (n > 1 && n != k) stop(
+    "the length of `theta` (", n, ") differs with the number of analyses (", k, ")!"
   )
-  if (tail(theta, 1) < 0) stop("final effect size of `theta` must be non-negative!")
+  if (last_(theta) < 0) stop(
+    "final effect size of `theta` (", deparse(theta), ") must be non-negative!"
+  )
 }
 
 #' Check the arguments `test_upper` and `test_lower`
 #'
 #' @param x Test upper or lower.
-#' @param K Number of total analysis.
+#' @param k Number of total analysis.
 #' @noRd
-check_test_upper <- function(x, K) {
-  check_test_ul(x, K, "test_upper")
+check_test_upper <- function(x, k) {
+  check_test_ul(x, k, "test_upper")
   # check that final test_upper value is TRUE
   if (!last_(x)) stop("last value of `test_upper` (", deparse(x), "), must be TRUE!")
 }
 
-check_test_lower <- function(x, K) check_test_ul(x, K, "test_lower")
+check_test_lower <- function(x, k) check_test_ul(x, k, "test_lower")
 
-check_test_ul <- function(x, K, name) {
+check_test_ul <- function(x, k, name) {
   # check length
-  if (!any(length(x) == c(1, K)))
+  if (!any(length(x) == c(1, k)))
     stop("`", name, "` must be of length 1 or same length as info!")
 }