--- a/R/check_arg.R
+++ b/R/check_arg.R
@@ -19,7 +19,7 @@
 # check if values in a vector are increasing
 check_increasing <- function(x, name = deparse(substitute(x)), first = TRUE) {
   # if first = TRUE, require x[1] to be positive, otherwise x[1] can be 0
-  d <- if (first) diff2(x) else diff(x)
+  d <- if (first) diff_one(x) else diff(x)
   if (any(d <= 0)) stop(
     "`", name, "` (", deparse(x), ") must be ",
     if (first) "positive and ", "strictly increasing!"