… NAMESPACE
… NEWS.md
… R/expected_event.R
… R/expected_time.R
… R/gsDesign2-package.R
… R/gs_design_ahr.R
… R/gs_design_combo.R
… R/gs_design_npe.R
… R/gs_design_wlr.R
… R/gs_info_ahr.R
… R/gs_info_wlr.R
… R/gs_power_npe.R
… R/pw_info.R
… R/pwe.R
… R/utility_combo.R
… R/utility_wlr.R
… R/utils.R
… man/event_diff.Rd
… man/expected_accrual.Rd
… man/fastlag.Rd
… man/get_combo_power.Rd
… man/gs_design_npe.Rd
… man/gs_power_npe.Rd
… tests/testthat/_snaps/independent_as_rtf/fixed_design_ahr.rtf
… tests/testthat/_snaps/independent_as_rtf/fixed_design_ahr_footnote.rtf
… tests/testthat/_snaps/independent_as_rtf/fixed_design_ahr_title.rtf
… tests/testthat/helper-old-version-eAccrual_.R
… tests/testthat/helper-old-version-eEvents_.R
… tests/testthat/helper-old-version-gs_design_npe_.R
… tests/testthat/helper-old-version-tEvents_.R