Close #527