@jdblischak 2025-04-16
> Yes, `to_integer()` is slow: [#527 (comment)](https://github.com/Merck/gsDesign2/issues/527#issuecomment-2797847696)

My apologies. I had seen your comment earlier but then forgot about it when I was performing the benchmarking yesterday.

> For now, `pw_info()` should be faster with this PR

Confirmed! This PR improves the performance, especially for the worst-case scenario (`max`).

```R
# main
git2r::commits(n = 1)[[1]]
## [6b8d5ac] 2025-04-14: Merge pull request #529 from Merck/gt-latex-table-position
library("gsDesign2")
library("microbenchmark")

benchmark <- function() gs_design_ahr() |> to_integer() |> summary() |> as_gt()
microbenchmark(benchmark())
## Unit: milliseconds
##         expr      min       lq     mean   median       uq      max neval
##  benchmark() 421.8117 447.9604 598.8056 463.4538 482.0702 2983.017   100

# PR
git2r::commits(n = 1)[[1]]
## [1d3675e] 2025-04-14: Merge branch 'main' into optimize-pw_info
library("gsDesign2")
library("microbenchmark")

benchmark <- function() gs_design_ahr() |> to_integer() |> summary() |> as_gt()
microbenchmark(benchmark())
## Unit: milliseconds
##         expr      min       lq     mean   median       uq      max neval
##  benchmark() 347.3618 374.5193 397.7408 386.4939 413.2976 581.9334   100
```