@jdblischak 2025-04-17
Great work @yihui! As suggested by @nanxstats, I benchmarked your latest commits with code generated by the app, and there are clear speed improvements already.

```R
# main
git2r::commits(n = 1)[[1]]
## [f08e563] 2025-04-16: Merge pull request #531 from Merck/519-developer-tests-of-gs_cp_npe
microbenchmark(benchmark())
## Unit: seconds
##         expr      min       lq     mean   median       uq      max neval
##  benchmark() 1.615996 1.734431 3.256798 1.789942 1.948333 11.20348   100

# PR
git2r::commits(n = 1)[[1]]
## [36a9e82] 2025-04-16: swap the inner and outer loops to save time on subsetting, i.e., subset by stratum first, then loop on total_duration, otherwise we may be repeatedly subsetting the data for each total_duration
microbenchmark(benchmark())
## Unit: seconds
##         expr      min       lq     mean   median       uq      max neval
##  benchmark() 1.199986 1.260796 1.708124 1.297672 1.346608 7.784739   100
```

<details><summary>Full benchmarking code</summary>
<p>

```R
git2r::commits(n = 1)[[1]]
library("gsDesign2")
library("microbenchmark")

benchmark <- function() {
  # Enrollment
  enroll_rate <- define_enroll_rate(duration = c(2, 2, 2, 18), rate = c(1, 2, 3, 4))

  # Failure rates
  time_point <- c(4, 6, 44)
  duration <- c(time_point[1], diff(time_point))
  control_rate <- log(2)/c(6, 6, 6)
  fail_rate <- define_fail_rate(duration, fail_rate = control_rate, dropout_rate = c(0.001, 0.001, 0.001), hr = c(1, 0.8, 0.6))

  gs_design_ahr(
    enroll_rate = enroll_rate,
    fail_rate = fail_rate,
    ratio = 1,
    alpha = 0.025,
    beta = 0.1,
    info_frac = (1:3)/3,
    analysis_time = 30,
    info_scale = "h0_h1_info",
    upper = "gs_spending_bound",
    upar = list(sf = "sfLDOF", total_spend = 0.025, param = NULL),
    lower = "gs_b",
    lpar = c(-Inf, -Inf, -Inf),
    binding = FALSE,
    h1_spending = TRUE,
    test_lower = FALSE
  ) |>
    to_integer() |>
    summary() |>
    gsDesign2::as_gt(
      title = "Efficacy bound only",
      subtitle = "alpha-spending",
      footnote = FALSE
    )
}
microbenchmark(benchmark())
```

</p>
</details>