@yihui 2025-04-17
I also noticed the time was cut by ~25% with this PR when running all tests (previously it took ~100s and now it takes ~75s). I feel these improvements are relatively minor and the bigger bottlenecks are elsewhere. I started with `pw_info()` since it "looked" relatively simple (but turned out to be a nontrivial mental challenge).

One thing that I want to discuss tomorrow is the input checks. I feel some checks are getting pedantic, even though the original goal might be to pursue rigor (following SOP), e.g.,

```
  if (!is.numeric(time)) {
    stop("gsDesign2: time in `expected_accrual()` must be a strictly increasing non-negative numeric vector!")
  }

  if (!is.logical(simple)) {
    stop("gsDesign2: simple in `expected_event()` must be logical!")
  }
```

This feels like forcing R to be a typed language when it isn't. If we have to do `if (is.numeric())`, `if (is.logical())`, etc, for _every single_ function no matter if the function is public or internal, we probably shouldn't use R in the first place. I think we should expect users to spend some time on the manual to learn the expected values of arguments, and we shouldn't take so much responsibility to guard against cases like users passing a character value to a numeric argument.

I heard that we wanted the error message to be informative when users input a wrong type of argument. Honestly, some of our error messages are not quite informative yet, such as the first example above (the error was not due to `time` not being "strictly increasing non-negative" but only not being "numeric").

Personally I'd recommend that we get rid of some of these checks, at least for lower-level functions, since these checks are performed repeatedly on the same values when a function is called in a loop, which is completely meaningless (the arguments have already been checked in the higher-level function) and only wastes time.