@jdblischak 2025-04-17
> Personally I'd recommend that we get rid of some of these checks, at least for lower-level functions, since these checks are performed repeatedly on the same values when a function is called in a loop, which is completely meaningless (the arguments have already been checked in the higher-level function) and only wastes time.

I completely agree. Ideally we want to perform input checks only at the beginning when the user has first provided the data, and then avoid them inside the more loop-intensive parts of the code base.

I do worry though since `pw_info()` is an exported function, and thus users may call it directly. Early assertions can avoid a lot of user troubleshooting.

One strategy I adopted in {workflowr} was to have separate external and internal functions, e.g. `fname()` and `fname_()`. In this hypothetical scenario, `fname()` is the documented, exported function that a user will call directly. It performs a ton of diagnostic checks with helpful error messages. Once the checks pass, it then calls `fname_()`. And then any time one of the package functions needs to internally call this function, they can bypass the checks by calling `fname_()`. This way user-input is always validated, but internal calls are as fast as possible.